Portmaster component: Updated $component_config variable usage, rd path variable names and reset logic in component_prepare.sh Removed redundant code in component_launcher.sh

+++ portmaster/component_prepare.sh
@@ -2,18 +2,19 @@
 
 # Setting component name and path based on the directory name
 component_name="$(basename "$(dirname "$(readlink -f "${BASH_SOURCE[0]}")")")"
-config="/app/retrodeck/components/$component_name/rd_config"
+component_config="/app/retrodeck/components/$component_name/rd_config"
 
-# TODO: MultiUser
-log i "----------------------"
-log i "Preparing $component_name"
-log i "----------------------"
+if [[ "$action" == "reset" ]]; then # Run reset-only commands
+  log i "----------------------"
+  log i "Preparing $component_name"
+  log i "----------------------"
 
-rm -rf "$XDG_DATA_HOME/PortMaster"
-unzip "/app/retrodeck/components/portmaster/PortMaster.zip" -d "$XDG_DATA_HOME/"
-cp -f "$XDG_DATA_HOME/PortMaster/retrodeck/PortMaster.txt" "$XDG_DATA_HOME/PortMaster/PortMaster.sh"
-chmod +x "$XDG_DATA_HOME/PortMaster/PortMaster.sh"
-rm -f "$rd_home_roms_path/portmaster/PortMaster.sh"
-install -Dm755 "$XDG_DATA_HOME/PortMaster/PortMaster.sh" "$rd_home_roms_path/portmaster/PortMaster.sh"
-create_dir "$XDG_DATA_HOME/PortMaster/config/"
-cp "$config/config.json" "$XDG_DATA_HOME/PortMaster/config/config.json"
+  rm -rf "$XDG_DATA_HOME/PortMaster"
+  unzip "/app/retrodeck/components/portmaster/PortMaster.zip" -d "$XDG_DATA_HOME/"
+  cp -f "$XDG_DATA_HOME/PortMaster/retrodeck/PortMaster.txt" "$XDG_DATA_HOME/PortMaster/PortMaster.sh"
+  chmod +x "$XDG_DATA_HOME/PortMaster/PortMaster.sh"
+  rm -f "$roms_path/portmaster/PortMaster.sh"
+  install -Dm755 "$XDG_DATA_HOME/PortMaster/PortMaster.sh" "$roms_path/portmaster/PortMaster.sh"
+  create_dir "$XDG_DATA_HOME/PortMaster/config/"
+  cp "$component_config/config.json" "$XDG_DATA_HOME/PortMaster/config/config.json"
+fi
